Create Main.cpp
Adds a C++ program that includes a function to calculate character offsets, throwing exceptions in error conditions. The function takes a start character and an offset integer as arguments, and returns the character that is offset positions away from start. If start is not a letter (A-Z, a-z), an invalidCharacterException is thrown. If the target is not a letter, an invalidRangeException is thrown. If the transition from start to target violates the upper- and lower-case transition rule, an invalidRangeException is thrown. The commit also includes a main function to test the character function with different arguments.