Create PartB-Main.cpp
Adds a C++ program that includes templated functions to perform math on varying numeric data types. The program defines two functions named "half" that accept a single argument, one for floating-point types and one for integer types. For floating-point types (float, double), the function returns half of the value passed in as an argument. For integer types (int), the function returns half the value, rounded according to usual rounding rules. The program also includes a driver in main to test these functions with different numeric types.